show in match

+++ app/src/main/java/com/miklesam/dotamatchsimulator/Hero.kt
@@ -1,3 +1,3 @@
 package com.miklesam.dotamanager.datamodels
 
-data class Hero(var positionX:Int,var positionY:Int)
+data class Hero(var positionX:Double,var positionY:Double)
